+++ plugin/src/_utils/index.js
@@ -2,4 +2,5 @@ export * from './colors.js';
 export * from './mappings.js';
 export * from './variants.js';
 export * from './utilities.js';
+export * from './getComponents.js';
 export * from './handlers/index.js';
